Add clipboard monitoring for info hashes.
For performance, this feature is activated only when activating the Add Torrent Link dialog.

Then it is kept active for 10 minutes, and if it detects a valid info hash in the clipboard,
it automatically adds it. On Linux, the selection buffer is also supported.

If 10 minutes pass without a valid info hash, the feature is automatically disabled by a QTimer.